Crop Stat/requirements.txt
numpy
pillow
reportlab
streamlit
tensorflow
python-dotenv
google-generativeai